fix(Affilations): crash when creating affiliation
